move duplicated call to constructCanvasMessage
